refactor: refactored the bulk upload code to implement an interface
AB#260
